requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.21.1
pandas==1.3.1
google-generativeai==0.3.1
jsonify==0.5
gunicorn==20.1.0 